test(mitm-addon): split usage reporting tests by ownership
